% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetSet.R
\docType{methods}
\name{AcousticEvent.accessors}
\alias{AcousticEvent.accessors}
\alias{settings}
\alias{settings,AcousticEvent-method}
\alias{settings<-}
\alias{settings<-,AcousticEvent-method}
\alias{localizations}
\alias{localizations,AcousticEvent-method}
\alias{localizations<-}
\alias{localizations<-,AcousticEvent-method}
\alias{detectors}
\alias{detectors,AcousticEvent-method}
\alias{detectors<-}
\alias{detectors<-,AcousticEvent-method}
\alias{visData}
\alias{visData,AcousticEvent-method}
\alias{visData<-}
\alias{visData<-,AcousticEvent-method}
\alias{behavior}
\alias{behavior,AcousticEvent-method}
\alias{behavior<-}
\alias{behavior<-,AcousticEvent-method}
\alias{erddap}
\alias{erddap,AcousticEvent-method}
\alias{erddap<-}
\alias{erddap<-,AcousticEvent-method}
\alias{species}
\alias{species,AcousticEvent-method}
\alias{species<-}
\alias{species<-,AcousticEvent-method}
\alias{files}
\alias{files,AcousticEvent-method}
\alias{files<-}
\alias{files<-,AcousticEvent-method}
\title{\code{AcousticEvent} accessors}
\usage{
settings(x, ...)

\S4method{settings}{AcousticEvent}(x, ...)

settings(x) <- value

\S4method{settings}{AcousticEvent}(x) <- value

localizations(x, ...)

\S4method{localizations}{AcousticEvent}(x, ...)

localizations(x) <- value

\S4method{localizations}{AcousticEvent}(x) <- value

detectors(x, ...)

\S4method{detectors}{AcousticEvent}(x, ...)

detectors(x) <- value

\S4method{detectors}{AcousticEvent}(x) <- value

visData(x, ...)

\S4method{visData}{AcousticEvent}(x, ...)

visData(x) <- value

\S4method{visData}{AcousticEvent}(x) <- value

behavior(x, ...)

\S4method{behavior}{AcousticEvent}(x, ...)

behavior(x) <- value

\S4method{behavior}{AcousticEvent}(x) <- value

erddap(x, ...)

\S4method{erddap}{AcousticEvent}(x, ...)

erddap(x) <- value

\S4method{erddap}{AcousticEvent}(x) <- value

species(x, ...)

\S4method{species}{AcousticEvent}(x, ...)

species(x) <- value

\S4method{species}{AcousticEvent}(x) <- value

files(x, ...)

\S4method{files}{AcousticEvent}(x, ...)

files(x) <- value

\S4method{files}{AcousticEvent}(x) <- value
}
\arguments{
\item{x}{a \linkS4class{AcousticEvent} object}

\item{\dots}{other arguments to pass to methods}

\item{value}{value to assign with accessor}
}
\value{
\describe{
  \item{settings}{a \linkS4class{DataSettings} object}
  \item{detectors}{a list of detector data frames}
  \item{localizations}{list of localizations}
  \item{visData}{a \linkS4class{VisObsData} object}
  \item{behavior}{behavior data}
  \item{erdda}{environmental data}
  \item{species}{list of species classifications}
  \item{files}{list of files used to create this object}
  }
}
\description{
Accessors for slots in \linkS4class{AcousticEvent} objects
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
