% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBinaries.R
\name{addBinaries}
\alias{addBinaries}
\title{Add Binaries to a PAMrSettings Object}
\usage{
addBinaries(prs, binfolder)
}
\arguments{
\item{prs}{a \linkS4class{PAMrSettings} object to add a database to}

\item{binfolder}{a folder of binaries to add}
}
\value{
the same \linkS4class{PAMrSettings} object as prs, with the binary
  files contained in \code{binfolder} added to the "binaries" slot
}
\description{
Adds a new function to the "function" slot in a PAMrSettings
  object.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
