% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCalibration.R
\name{addCalibration}
\alias{addCalibration}
\alias{applyCalibration}
\title{Add a Calibration File to a PAMrSettings Object}
\usage{
addCalibration(prs, calFile = NULL, module = "ClickDetector", calName = NULL)

applyCalibration(prs, module = "ClickDetector")
}
\arguments{
\item{prs}{a \linkS4class{PAMrSettings} object to add a database to}

\item{calFile}{a calibration file name. Must be csv format with two columns.
The first column must be the frequency, and the second column must be the
sensitivity (in dB), and the columns should be labeled \code{Frequency}
and \code{Sensitivity}. Can also be supplied as a dataframe in which
case the \code{calName} argument should also be set}

\item{module}{the Pamguard module type this calibration should be applied to,
for now this is only for ClickDetector modules. This is left as an option
for future-proofing purposes but should not be needed.}

\item{calName}{the name to assign to the calibration function, defaults to
the file name and only needs to be set if supplying a dataframe instead of
a csv file}
}
\value{
the same \linkS4class{PAMrSettings} object as prs, with the calibration
  function added to the \code{calibration} slot.
}
\description{
Adds a new calibration function to the "calibration" slot
  of a PAMrSettings object.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
