% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFunction.R
\name{addFunction}
\alias{addFunction}
\title{Add a Function to a PAMrSettings Object}
\usage{
addFunction(prs, fun, module)
}
\arguments{
\item{prs}{a \linkS4class{PAMrSettings} object to add a function to}

\item{fun}{function to add}

\item{module}{PamGuard module output this function should act on}
}
\value{
the same \linkS4class{PAMrSettings} object as prs, with the function
  \code{fun} added to the "functions" slot
}
\description{
Adds a new function to the "function" slot in a PAMrSettings
  object.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
