% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Cruise-class}
\alias{Cruise-class}
\title{\code{Cruise} Class}
\description{
An S4 class storing acoustic data from an entire cruise
}
\section{Slots}{

\describe{
\item{\code{folders}}{a list of folders containing the cruise data}

\item{\code{gpsData}}{a data frame of gps coordinates for the entire cruise}

\item{\code{acousticEvents}}{a list of \linkS4class{AcousticEvent} objects with
detections from the cruise}

\item{\code{detectorSettings}}{a named list of settings for the detectors. Names should
match the detectors found in the \code{acousticEvents} list}

\item{\code{localizationSettings}}{a named list of settings for each type of localization done}

\item{\code{effort}}{something about effort}
}}

\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
