% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetSet.R
\docType{methods}
\name{Cruise.accessors}
\alias{Cruise.accessors}
\alias{folders}
\alias{folders,Cruise-method}
\alias{folders<-}
\alias{folders<-,Cruise-method}
\alias{gpsData}
\alias{gpsData,Cruise-method}
\alias{gpsData<-}
\alias{gpsData<-,Cruise-method}
\alias{acousticEvents}
\alias{acousticEvents,Cruise-method}
\alias{acousticEvents<-}
\alias{acousticEvents<-,Cruise-method}
\alias{detectorSettings}
\alias{detectorSettings,Cruise-method}
\alias{detectorSettings<-}
\alias{detectorSettings<-,Cruise-method}
\alias{localizationSettings}
\alias{localizationSettings,Cruise-method}
\alias{localizationSettings<-}
\alias{localizationSettings<-,Cruise-method}
\alias{effort}
\alias{effort,Cruise-method}
\alias{effort<-}
\alias{effort<-,Cruise-method}
\title{\code{Cruise} accessors}
\usage{
folders(x, ...)

\S4method{folders}{Cruise}(x, ...)

folders(x) <- value

\S4method{folders}{Cruise}(x) <- value

gpsData(x, ...)

\S4method{gpsData}{Cruise}(x, ...)

gpsData(x) <- value

\S4method{gpsData}{Cruise}(x) <- value

acousticEvents(x, ...)

\S4method{acousticEvents}{Cruise}(x, ...)

acousticEvents(x) <- value

\S4method{acousticEvents}{Cruise}(x) <- value

detectorSettings(x, ...)

\S4method{detectorSettings}{Cruise}(x, ...)

detectorSettings(x) <- value

\S4method{detectorSettings}{Cruise}(x) <- value

localizationSettings(x, ...)

\S4method{localizationSettings}{Cruise}(x, ...)

localizationSettings(x) <- value

\S4method{localizationSettings}{Cruise}(x) <- value

effort(x, ...)

\S4method{effort}{Cruise}(x, ...)

effort(x) <- value

\S4method{effort}{Cruise}(x) <- value
}
\arguments{
\item{x}{a \linkS4class{Cruise} object}

\item{\dots}{other arguments to pass to methods}

\item{value}{value to assign with accessor}
}
\value{
\describe{
  \item{folders}{a list of folders}
  \item{gpsData}{a dataframe containing gps data}
  \item{acousticEvents}{a list of \linkS4class{AcousticEvent} objects}
  \item{detectorSettings}{a named list of settings for each detector used}
  \item{localizationSettings}{a named list of settings for each localization}
  \item{effort}{something about effort?}
  }
}
\description{
Accessors for slots in \linkS4class{Cruise} objects
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
