% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{DataSettings-class}
\alias{DataSettings-class}
\title{An S4 class to store data collection settings. Possible inclusions are
hydrophone sensitivity, sample rate, sound card, etc.}
\description{
An S4 class to store data collection settings. Possible inclusions are
hydrophone sensitivity, sample rate, sound card, etc.
}
\section{Slots}{

\describe{
\item{\code{sampleRate}}{the sample rate data was recorded at.}

\item{\code{soundSource}}{the source of the recorded sound - sound card, recording
system, or sound file}
}}

