% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGetSet.R
\docType{methods}
\name{DataSettings.accessors}
\alias{DataSettings.accessors}
\alias{sampleRate}
\alias{sampleRate,DataSettings-method}
\alias{sampleRate<-}
\alias{sampleRate<-,DataSettings-method}
\alias{soundSource}
\alias{soundSource,DataSettings-method}
\alias{soundSource<-}
\alias{soundSource<-,DataSettings-method}
\title{\code{DataSettings} accessors}
\usage{
sampleRate(x, ...)

\S4method{sampleRate}{DataSettings}(x, ...)

sampleRate(x) <- value

\S4method{sampleRate}{DataSettings}(x) <- value

soundSource(x, ...)

\S4method{soundSource}{DataSettings}(x, ...)

soundSource(x) <- value

\S4method{soundSource}{DataSettings}(x) <- value
}
\arguments{
\item{x}{a \linkS4class{DataSettings} object}

\item{\dots}{other arguments to pass to methods}

\item{value}{value to assign with accessor}
}
\value{
\describe{
  \item{sampleRate}{sample rate in Hz of the sound}
  \item{soundSource}{source of the sound}
  }
}
\description{
Accessors for slots in \linkS4class{DataSettings} objects
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
