% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_banter.R
\name{export_banter}
\alias{export_banter}
\title{Export a List of AcousticEvent Ojbects for Banter Model}
\usage{
export_banter(eventList)
}
\arguments{
\item{eventList}{a list of \linkS4class{AcousticEvent} objects.}
}
\value{
a list with two items, \code{events} and \code{detections}.
  \code{events} is a dataframe with two columns. \code{event.id} is a
  unique identifier for each event, taken from the names of the event
  list. \code{species} is the species classification, taken from the
  \code{species} slot labelled \code{id}. \code{detections} is a list
  of data frames containing all the detections and measurements. There is
  one list for each unique detector type found in the \code{detectors} slots
  of \code{eventList}. The data frames will only have columns with class
  \code{numeric}, \code{integer}, \code{factor}, or \code{logical}, and
  will also have columns named \code{UID}, \code{Id}, \code{parentUID},
  \code{sampleRate}, and \code{Channel} removed so that these are not treated
  as parameters for the banter random forest model. The dataframes will also
  have columns \code{event.id} and \code{call.id} added.
}
\description{
Formats a list of AcousticEvent objects into the structure
  needed to run a banter model.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
