% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBinaryData.R
\name{getBinaryData}
\alias{getBinaryData}
\title{Get Binary Data for Detections from an Acoustic Event}
\usage{
getBinaryData(acEv, UID)
}
\arguments{
\item{acEv}{a \linkS4class{AcousticEvent} object}

\item{UID}{the UID(s) of the individual detections to fetch the binary
data for}
}
\value{
a list of \code{PamBinary} objects for each \code{UID}
}
\description{
Fetches matching binary data from a single or multiple
  detections in an \linkS4class{AcousticEvent} object
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
