% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPgDetections.R
\name{getPgDetections}
\alias{getPgDetections}
\title{Load and Process Detections from Pamguard}
\usage{
getPgDetections(prs, mode = c("db", "all", "time"), id = NULL,
  sampleRate = NULL, grouping = c("event", "detGroup"),
  format = "\%Y-\%m-\%d \%H:\%M:\%OS", ...)
}
\arguments{
\item{prs}{a \linkS4class{PAMrSettings} object containing the databases,
binaries, and functions to use for processing data.
See \code{\link[PAMr]{PAMrSettings}}}

\item{mode}{selector for how to organize your data in to events. \code{db}
will organize by events based on tables in the databases, \code{all} will
put all detections in the binaries into one single event, and \code{time}
will organize into events based on timestamps provided in \code{grouping}.}

\item{id}{an event name or id for this event, only used for \code{mode='all'}}

\item{sampleRate}{the sample rate of the data. If reading detections from
a database this will be read in from the SoundAcquisition table, if
\code{mode} is \code{all} or \code{time} this must be specified.}

\item{grouping}{if \code{mode} is \code{db}, the table to group events by.
Either \code{event} to use the OfflineEvents table, or \code{detGroup} to
use the detection group localiser module groups. If \code{mode} is
\code{time}, this should be a data frame with four columns and 1 row
for each separate event. \code{start} and \code{end} should specify the
start and end time of the event and must be in UTC. \code{id} should
specify a unique id for each event, and \code{species} should provide a
species ID if it is available (this column is optional). This can also
be a character with the file path to a csv file with these columns.}

\item{format}{the date format for the \code{start} and \code{end} columns
in \code{grouping} if it is a csv. Times are assumed to be UTC. See
?strptime for details.}

\item{\dots}{additional arguments to pass onto to different methods}
}
\value{
a list of \linkS4class{AcousticEvent} objects. If reading from a
  database, there will be a separate AcousticEvent for each event found
  by the grouping method chosen. If just reading all detections in a set
  of binary files, all the detections will be in a single AcousticEvent.
}
\description{
Loads in acoustic detection data that has been run thorugh
  Pamguard. Uses the binary file output from Pamguard and (optionally)
  the database used. If a database is used then only detections found
  in that database will be processed, and they will be grouped by either
  the 'OfflineEvents' table or the 'Detection Group Localiser' modulef
  groups. If no database is provided, all detections in all of the
  provided binary files will be processed, this can take an extremely
  long time. This will also apply a set of functions to the binaray data
  loaded. Can also provide a database or csv file to separate detections
  from binaries into events by start and end times.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
