% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamr-package.R
\docType{package}
\name{pamr-package}
\alias{pamr}
\alias{pamr-package}
\alias{pamr.cube.root}
\alias{pamr.pairscore}
\alias{pamr.pvalue.survival}
\alias{pamr.score.to.class1}
\alias{pamr.score.to.class2}
\alias{print.nsc}
\alias{print.nsccv}
\alias{print.pamrcved}
\alias{print.pamrtrained}
\title{Prediction Analysis of Microarrays}
\description{
Functions for training and predicting using shrunken
centroids. While originally developed in the context of
microarray data analysis (\url{doi:10.1073/pnas.082099299}), the
method is broadly applicable.
}
\author{
\strong{Maintainer}: Balasubramanian Narasimhan \email{naras@stanford.edu}

Authors:
\itemize{
  \item Trevor Hastie \email{hastie@stanford.edu}
  \item Rob Tibshirani \email{tibs@stanford.edu}
  \item Gilbert Chu
}

}
\keyword{internal}
