% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBinaries.R
\name{removeBinaries}
\alias{removeBinaries}
\title{Remove Binaries from a PAMrSettings Object}
\usage{
removeBinaries(prs, index = NULL)
}
\arguments{
\item{prs}{a \linkS4class{PAMrSettings} object to remove binaries from}

\item{index}{index indicating which binary folders to remove. Can be a vector
if you want to remove multiple folders. If missing user is prompted to
select a folder from a list, will only show up to the first 20. You can
easily remove all of the folders with a large index like \code{1:1000}}
}
\value{
the same \linkS4class{PAMrSettings} object as prs, with the binary
  folders and files associated with those folders removed from the "binaries"
  slot.
}
\description{
Remove a binary folder and associated files from the "binaries"
  slot in a PAMrSettings object.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
