% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setIdSlot.R
\name{setIdSlot}
\alias{setIdSlot}
\title{Temp Function to Assign Id Slot}
\usage{
setIdSlot(evList)
}
\arguments{
\item{evList}{a named list of \linkS4class{AcousticEvent} objects}
}
\value{
the same list of \linkS4class{AcousticEvent} objects, but with the
  id slot added and set to the names of the list
}
\description{
Uses the names of a list to set the Id slot of AcousticEvent
  objects, only used to update to v0.6.0, not needed after
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
