% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSpecies.R
\name{setSpecies}
\alias{setSpecies}
\title{Set the Species Classification of an Acoustic Event}
\usage{
setSpecies(acev, type = "id", method = c("pamguard", "manual"), value)
}
\arguments{
\item{acev}{a \linkS4class{AcousticEvent} object, or list of objects}

\item{type}{the type of classification to set, this is just a label within
the \code{species} slot}

\item{method}{the method for assigning species to an event. Currently supports
\code{pamguard}, which will use the 'eventType' or 'Text_Annotation' column
to assign species, or \code{manual} which will use \code{value} to assign
species manually.}

\item{value}{optional argument required if \code{method} is set to manual.
Can either be a single value to assign to all events, or if assigning to
a list a vector with length equal to the list. Can also be a dataframe
with columns \code{event} and \code{species}, in which case species will
be matched to corresponding event names instead of just relying on the
order. If using this, please note the prefix OE or DGL present on most
event numbers (see the \code{id} slot of your events).}
}
\value{
the same object as \code{acev}, with species identifications assigned
  as an item named \code{type} in the \code{species} slot
}
\description{
Sets the \code{species} slot of an \linkS4class{AcousticEvent}
  object or list of objects
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
