% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardClickCalcs.R
\name{standardClickCalcs}
\alias{standardClickCalcs}
\title{Calculate a Set of Measurements for Clicks}
\usage{
standardClickCalcs(data, calibration, highpass_khz = 10,
  winLen_sec = 0.0025)
}
\arguments{
\item{data}{a list that must have 'wave' containing the wave form as a
matrix with a separate column for each channel, and 'sampleRate' the
sample rate of the data. Data can also be a \code{Wave} class 
object, like one created by \code{\link[tuneR]{Wave}}.}

\item{calibration}{a calibration function to apply to the spectrum, must be
a gam. If missing no calibration will be applied (not recommended).}

\item{highpass_khz}{frequency in khz of highpass filter to apply}

\item{winLen_sec}{length in seconds of fft window. The click wave is first
shortened to this number of samples around the peak of the wave,
removing a lot of the noise around the click. Following approach of
JB/EG/MS.}
}
\value{
A data frame with one row for each channel of click waveform.
  Calculates approximate noise level and click duration from the
  TK energy (Soldevilla JASA17), up to 3 highest peak frequencies and
  the 'troughs' between them (see \code{\link[PAMmisc]{peakTrough}}), and the 3
  and 10dB bandwidth levels and 'Q' value (see \code{\link[seewave]{Q}}).
}
\description{
Calculate a set of "standard" measurements for odontocete clicks.
  Most of calculations following approach of Baumann-Pickering / Soldevilla
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
