% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{VisObsData-class}
\alias{VisObsData-class}
\title{\code{VisObsData} Class}
\description{
An S4 class storing visual obsever data (UNFINISHED)
}
\section{Slots}{

\describe{
\item{\code{detectionTime}}{detection time}

\item{\code{speciesId}}{species id}

\item{\code{groupSizeEst}}{group size estimate}

\item{\code{effortStatus}}{effort status}
}}

\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
