% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBinaries.R
\name{addBinaries}
\alias{addBinaries}
\title{Add Binaries to a PAMrSettings Object}
\usage{
addBinaries(prs, binFolder = NULL)
}
\arguments{
\item{prs}{a \linkS4class{PAMrSettings} object to add a database to}

\item{binFolder}{a folder of binaries to add}
}
\value{
the same \linkS4class{PAMrSettings} object as prs, with the binary
  files contained in \code{binFolder} added to the "binaries" slot. Only
  binary files for Click Detector and WhistlesMoans modules will be added,
  since these are the only types PAMr currently knows how to process
  (last updated v 0.7.0)
}
\description{
Adds a new function to the "function" slot in a PAMrSettings
  object.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
