% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDatabase.R
\name{addDatabase}
\alias{addDatabase}
\title{Add a Database to a PAMrSettings Object}
\usage{
addDatabase(prs, db = NULL)
}
\arguments{
\item{prs}{a \linkS4class{PAMrSettings} object to add a database to}

\item{db}{a database to add}
}
\value{
the same \linkS4class{PAMrSettings} object as prs, with the database
  \code{db} added to the "db" slot
}
\description{
Adds a new function to the "function" slot in a PAMrSettings
  object.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
