% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPgDetections.R
\name{getPgDetections}
\alias{getPgDetections}
\alias{getPgDetectionsAll}
\alias{getPgDetectionsDb}
\title{Load and Process Detections from Pamguard}
\usage{
getPgDetections(prs, ...)

getPgDetectionsAll(prs, sampleRate)

getPgDetectionsDb(prs, grouping = c("event", "detGroup"))
}
\arguments{
\item{prs}{a \linkS4class{PAMrSettings} object containing the databases,
binaries, and functions to use for processing data.
See \code{\link[PAMr]{PAMrSettings}}}

\item{\dots}{other parameters to be passed on to other methods}

\item{sampleRate}{the sample rate of the data. If reading detections from
a database this will be read in from the SoundAcquisition table, if
loading all binary files this must be specified.}

\item{grouping}{if reading from a database, the table to group events by.
Either \code{event} to use the OfflineEvents table, or \code{detGroup} to
use the detection group localiser module groups.}
}
\value{
a list of \linkS4class{AcousticEvent} objects. If reading from a
  database, there will be a separate AcousticEvent for each event found
  by the grouping method chosen. If just reading all detections in a set
  of binary files, all the detections will be in a single AcousticEvent.
}
\description{
Loads in acoustic detection data that has been run thorugh
  Pamguard. Uses the binary file output from Pamguard and (optionally)
  the database used. If a database is used then only detections found
  in that database will be processed, and they will be grouped by either
  the 'OfflineEvents' table or the 'Detection Group Localiser' module
  groups. If no database is provided, all detections in all of the
  provided binary files will be processed, this can take an extremely
  long time. This will also apply a set of functions to the binaray data
  loaded. \code{getPgDetections} is just a wrapper that will call
  either \code{getPgDetectionsAll} or \code{getPgDetectionsDb} based
  on whether or not a database file is provided.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
