\name{khan}
\alias{khan}
\title{Khan microarray data}
\description{
 The \code{khan} data frame has 2309 rows and 65 columns.
These are one of the datasets data used in the Tibshirani et al paper in PNAS
on nearest shrunken centroids.
}
\format{
}
\details{
The first row contains the sample labels. The first two columns of gene ids and names. 
The remaining values of the matrix are gene expression values.
}
\source{
  }
\references{
}
\keyword{datasets}

