% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khan.R
\name{khan}
\alias{khan}
\title{Khan microarray data}
\description{
The \code{khan} data frame has 2308 rows and 65 columns. These are one of
the datasets data used in the Tibshirani et al paper in PNAS on nearest
shrunken centroids.
}
\details{
The first two columns of gene ids and names and the remaining
columns are gene expression values for 63 samples. An attribute
\code{cancer_type} contains the cancer type for each sample.
}
\keyword{datasets}
