\name{pamr.internal}
\alias{pamr.internal}

\title{Internal functions used by pamr}
\alias{balanced.folds}
\alias{softmax}
\alias{descendants}
\alias{diag.disc}
\alias{enlist}
\alias{error.bars}
\alias{error.nsc}
\alias{misreg.simple}
\alias{nnmiss}
\alias{nsc}
\alias{nsccv}
\alias{pamr.cube.root}
\alias{pamr.pairscore}
\alias{pamr.predictmany}
\alias{pamr.xl.compute.confusion}
\alias{pamr.xl.compute.cv.confusion}
\alias{pamr.xl.compute.offset}
\alias{pamr.xl.derive.adjusted.prior}
\alias{pamr.xl.get.default.training.parameters}
\alias{pamr.xl.get.offset}
\alias{pamr.xl.get.sample.prior}
\alias{pamr.xl.get.uniform.prior}
\alias{pamr.xl.is.a.subset}
\alias{pamr.xl.listgenes.compute}
\alias{pamr.xl.plot.test.probs.compute}
\alias{pamr.xl.plot.training.error.compute}
\alias{pamr.xl.plotcen.compute}
\alias{pamr.xl.plotcv.compute}
\alias{pamr.xl.plotcvprob.compute}
\alias{pamr.xl.predict.test.class}
\alias{pamr.xl.predict.test.class.only}
\alias{pamr.xl.predict.test.probs}
\alias{pamr.xl.process.data}
\alias{pamr.xl.test.data.impute}
\alias{pamr.xl.test.errors.compute}
\alias{pamr.xl.transform.class.labels}
\alias{pamr.xl.transform.data}
\alias{pamr.xl.transform.test.data}
\alias{permute.rows}
\alias{print.nsc}
\alias{print.nsccv}
\alias{roc.nsc}
\alias{soft.shrink}
\alias{softmax}
}
\description{Internal functions used by pamr}
   
\usage{}

\arguments{}
    
\details{
These
functions are internal to the package and not meant to be called by
users. 
}

\value{}

\references{}


\author{ Trevor Hastie, Robert Tibshirani, Balasubramanian Narasimhan, and Gilbert Chu  }


\keyword{ }

