% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCalibration.R
\name{removeCalibration}
\alias{removeCalibration}
\title{Remove a Calibration Function from a PAMrSettings Object}
\usage{
removeCalibration(prs, index = NULL, module = "ClickDetector")
}
\arguments{
\item{prs}{a \linkS4class{PAMrSettings} object to remove a calibration from}

\item{index}{index of the calibration function to remove. If \code{NULL}, user
will be prompted to select from a list. This can also be a vector to remove
multiple calibration functions at once.}

\item{module}{the module of the calibration function to remove, currently
not needed}
}
\value{
the same \linkS4class{PAMrSettings} object as prs, with the calibration
  function removed from the "calibration" slot
}
\description{
Remove a calibration function from the "calibration"
  slot of a PAMrSettings object
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
