% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showWaveform.R
\name{showWaveform}
\alias{showWaveform}
\alias{showSpectrogram}
\alias{showWigner}
\title{Show Graphical Representations of Waveforms}
\usage{
showWaveform(acEv, UID)

showSpectrogram(acEv, UID, ...)

showWigner(acEv, UID, ...)
}
\arguments{
\item{acEv}{a \linkS4class{AcousticEvent} object}

\item{UID}{the UID(s) of the individual detections to fetch the binary 
data for}

\item{\dots}{other arguments to pass to the spectrogram or wigner functions}
}
\value{
Nothing, just shows plots for every channel of the waveform for
  each UID provided
}
\description{
Fetches matching binary data from a single or multiple
  detections in an \linkS4class{AcousticEvent} object, then plot
  the resulting data
}
\details{
The \code{showSpectrogram} function uses the function 
  \code{\link[signal]{specgram}} to plot the spectrogram, see this function
  for plotting options. The \code{showWigner} function uses the function
  \code{\link[PAMmisc]{wignerTransform}} to plot the Wigner-Ville transform,
  see this function for options.
}
\author{
Taiki Sakai \email{taiki.sakai@noaa.gov}
}
